\name{dq_write_json}
\alias{dq_write_json}
\title{Write export-shaped data to JSON}
\usage{
dq_write_json(x, path)
}
\arguments{
\item{x}{Export-shaped data list.}
\item{path}{Destination file path.}
}
\value{The \code{path}, invisibly.}
\description{Write export-shaped data to disk in JSON format.}
